+++ CHANGELOG.md
@@ -5,8 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Possibility to add a **Table of Content** [#14]
+
 ### Changed
 - Update doc
+- CSS correction 
+- Add **Table of Content** in the CSS Squeleton
+
+---
 
 ## [0.4.0] - 2021-08-11
 ### Added 
@@ -20,7 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Remove **page** option
     - Add **output** option to generate PDF [#6]
 - CSS correction [#13]
-- Add **Table of Content** in the CSS Squeleton
+
+---
 
 ## [0.3.0] - 2021-08-08
 ### Added
@@ -31,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Documentation
 - CLI command to allow multiple files in entry [#3]
+
+---
 
 ## [0.2.0] - 2021-08-07
 ### Added
@@ -52,6 +62,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - code blocks
     - checkbox list
 - Default CSS style
+
+---
 
 ## [0.1.0] - 2021-07-31
 ### Added